+++ frontend/lib/app/app.dart
@@ -1,11 +1,11 @@
 import 'package:app_ui/app_ui.dart';
-import 'package:fieldresearch/screens/adm_page/create_form_view/create_form_view.dart';
-import 'package:fieldresearch/screens/adm_page/home_adm_view/home_adm_view.dart';
-import 'package:fieldresearch/screens/adm_page/users_adm_view/users_adm_view.dart';
-import 'package:fieldresearch/screens/researcher_page/reseacher_home_view.dart';
-import 'package:fieldresearch/screens/sign_in_page/bloc/sign_in_bloc.dart';
-import 'package:fieldresearch/screens/sign_in_page/view/sign_in_view.dart';
-import 'package:fieldresearch/screens/sign_up_page/view/sign_up_view.dart';
+import 'package:fieldresearch/create_form_page/create_form_view.dart';
+import 'package:fieldresearch/home_adm_page/home_adm_view.dart';
+import 'package:fieldresearch/users_adm_page/users_adm_view.dart';
+import 'package:fieldresearch/researcher_page/reseacher_home_view.dart';
+import 'package:fieldresearch/sign_in_page/bloc/sign_in_bloc.dart';
+import 'package:fieldresearch/sign_in_page/view/sign_in_view.dart';
+import 'package:fieldresearch/sign_up_page/view/sign_up_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
