+++ frontend/lib/create_form_page/widgets/int_type_data.dart
@@ -1,4 +1,3 @@
-import 'package:fieldresearch/utils/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -28,7 +27,7 @@ class _IntTypeDataState extends State<IntTypeData> {
                     decoration: InputDecoration(
                       contentPadding: EdgeInsets.symmetric(vertical: 14.h),
                       filled: true,
-                      fillColor: fillFormColor,
+                      fillColor: Colors.amber,
                       border: OutlineInputBorder(
                         borderRadius: BorderRadius.circular(3.r),
                         borderSide: BorderSide.none,
@@ -55,7 +54,7 @@ class _IntTypeDataState extends State<IntTypeData> {
                     decoration: InputDecoration(
                       contentPadding: EdgeInsets.symmetric(vertical: 14.h),
                       filled: true,
-                      fillColor: fillFormColor,
+                      fillColor: Colors.amber,
                       border: OutlineInputBorder(
                         borderRadius: BorderRadius.circular(3.r),
                         borderSide: BorderSide.none,
