+++ frontend/lib/researcher_page/reseacher_home_view.dart
@@ -1,5 +1,5 @@
 import 'package:app_ui/app_ui.dart';
-import 'package:fieldresearch/screens/researcher_page/researches_view.dart';
+import 'package:fieldresearch/researcher_page/researches_view.dart';
 import 'package:flutter/material.dart';
 
 import 'package:flutter_bloc/flutter_bloc.dart';
