+++ frontend/lib/researcher_page/researches_view.dart
