--- frontend/lib/utils/utils.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-
-const backGroundColor = Color(0xFF2D323D);
-const fillFormColor = Color(0xFF1B1F27);
-const textColorForm = Color(0xFF8F96A4);
-const buttonColor = Color(0xFF1B1F27);
