+++ frontend/lib/widgets/button_adm.dart
@@ -1,4 +1,3 @@
-import 'package:fieldresearch/utils/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
@@ -27,7 +26,7 @@ class MyButtonAdm extends StatelessWidget {
           ),
         ),
         backgroundColor: const MaterialStatePropertyAll(
-          buttonColor,
+          Colors.grey,
         ),
       ),
       onPressed: onPressed,
