+++ frontend/packages/user_repository/lib/src/spring_user_repository.dart
@@ -1,9 +1,7 @@
 import 'dart:convert';
 import 'dart:developer';
-
 import 'package:app_client/app_client.dart';
 import 'package:dartz/dartz.dart';
-import 'package:http/http.dart';
 import 'package:user_repository/src/models/user_model.dart';
 import 'user_repo.dart';
 
@@ -79,7 +77,7 @@ class SpringUserRepository implements IUserRepository {
         'email': email,
         'username': username,
         'password': password,
-        'role': "ADMIN",
+        'role': "USER",
       },
     );
     try {
